refactor(mcp): move MCP server from ui/backend/mcp to evidencelab_mcp
The MCP server is an independent service, not part of the UI.
Move to project root as evidencelab_mcp/ (not mcp/ to avoid
collision with the mcp pip package).

Updated all imports and docker-compose command.

Tested: all 26 unit tests pass.
