+++ evidencelab_mcp/tools/document.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 from typing import Any, Dict, Optional
 
-from ui.backend.mcp.schemas import MCPDocumentResponse
+from evidencelab_mcp.schemas import MCPDocumentResponse
 
 logger = logging.getLogger(__name__)
